test: duplicate pages for edge instead of using next.patch
