+++ test/e2e/app-dir/next-after-app/app/edge/[id]/dynamic/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../../nodejs/[id]/dynamic/page'
