+++ test/e2e/app-dir/next-after-app/app/edge/[id]/layout.js
@@ -1,0 +1,1 @@
+export { default } from '../../nodejs/[id]/layout'
