+++ test/e2e/app-dir/next-after-app/app/edge/[id]/with-action/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../../nodejs/[id]/with-action/page'
