+++ test/e2e/app-dir/next-after-app/app/edge/[id]/with-metadata/page.js
@@ -1,0 +1,4 @@
+export {
+  default,
+  generateMetadata,
+} from '../../../nodejs/[id]/with-metadata/page'
