+++ test/e2e/app-dir/next-after-app/app/edge/delay/page.js
@@ -1,0 +1,3 @@
+export { default } from '../../nodejs/delay/page'
+
+export const dynamic = 'force-dynamic'
