+++ test/e2e/app-dir/next-after-app/app/edge/interrupted/calls-not-found/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../../nodejs/interrupted/calls-not-found/page'
