+++ test/e2e/app-dir/next-after-app/app/edge/interrupted/calls-redirect/page.js
@@ -1,0 +1,5 @@
+import { createPage } from '../../../nodejs/interrupted/calls-redirect/page'
+
+// NOTE: this page is forked from /nodejs
+
+export default createPage('/edge')
