+++ test/e2e/app-dir/next-after-app/app/edge/interrupted/redirect-target/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../../nodejs/interrupted/redirect-target/page'
