+++ test/e2e/app-dir/next-after-app/app/edge/interrupted/throws-error/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../../nodejs/interrupted/throws-error/page'
