+++ test/e2e/app-dir/next-after-app/app/edge/invalid-in-client/page.js
@@ -1,0 +1,6 @@
+// (replaced in tests)
+// 'use client'
+
+export { default } from '../../nodejs/invalid-in-client/page'
+
+export const dynamic = 'force-dynamic'
