+++ test/e2e/app-dir/next-after-app/app/edge/layout.js
@@ -1,0 +1,3 @@
+export { default } from '../nodejs/layout'
+
+export const runtime = 'edge'
