+++ test/e2e/app-dir/next-after-app/app/edge/middleware/redirect/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../../nodejs/middleware/redirect/page'
