+++ test/e2e/app-dir/next-after-app/app/edge/nested-after/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../nodejs/nested-after/page'
