+++ test/e2e/app-dir/next-after-app/app/edge/provided-request-context/page.js
@@ -1,0 +1,1 @@
+export { default } from '../../nodejs/provided-request-context/page'
