+++ test/e2e/app-dir/next-after-app/app/edge/route/route.js
@@ -1,0 +1,4 @@
+export { GET } from '../../nodejs/route/route'
+
+export const runtime = 'edge'
+export const dynamic = 'force-dynamic'
