+++ test/e2e/app-dir/next-after-app/app/edge/static/page.js
@@ -1,0 +1,4 @@
+export { default } from '../../nodejs/static/page'
+
+// (patched in tests)
+// export const dynamic = 'REPLACE_ME'
