--- test/e2e/app-dir/next-after-app/app/layout.js
@@ -1,13 +1,0 @@
-// (patched in tests)
-// export const runtime = 'REPLACE_ME'
-
-export default function AppLayout({ children }) {
-  return (
-    <html>
-      <head>
-        <title>after</title>
-      </head>
-      <body>{children}</body>
-    </html>
-  )
-}
