+++ test/e2e/app-dir/next-after-app/app/nodejs/[id]/layout.js
@@ -1,0 +1,9 @@
+import { unstable_after as after } from 'next/server'
+import { cliLog } from '../../../utils/log'
+
+export default function Layout({ children }) {
+  after(async () => {
+    cliLog({ source: '[layout] /[id]' })
+  })
+  return <>{children}</>
+}
