+++ test/e2e/app-dir/next-after-app/app/nodejs/interrupted/calls-not-found/page.js
@@ -1,0 +1,12 @@
+import { notFound } from 'next/navigation'
+import { unstable_after as after } from 'next/server'
+import { cliLog } from '../../../../utils/log'
+
+export default function Page() {
+  after(() => {
+    cliLog({
+      source: '[page] /interrupted/calls-not-found',
+    })
+  })
+  notFound()
+}
