+++ test/e2e/app-dir/next-after-app/app/nodejs/interrupted/calls-redirect/page.js
@@ -1,0 +1,19 @@
+import { unstable_after as after } from 'next/server'
+import { redirect } from 'next/navigation'
+import { cliLog } from '../../../../utils/log'
+
+// NOTE: this page is forked in /edge
+
+export function createPage(pathPrefix) {
+  return function Page() {
+    after(() => {
+      cliLog({
+        source: '[page] /interrupted/calls-redirect',
+      })
+    })
+
+    redirect(pathPrefix + '/interrupted/redirect-target')
+  }
+}
+
+export default createPage('/nodejs')
