+++ test/e2e/app-dir/next-after-app/app/nodejs/interrupted/throws-error/page.js
@@ -1,0 +1,11 @@
+import { unstable_after as after } from 'next/server'
+import { cliLog } from '../../../../utils/log'
+
+export default function Page() {
+  after(() => {
+    cliLog({
+      source: '[page] /interrupted/throws-error',
+    })
+  })
+  throw new Error('User error')
+}
