+++ test/e2e/app-dir/next-after-app/app/nodejs/layout.js
@@ -1,0 +1,12 @@
+export const runtime = 'nodejs'
+
+export default function AppLayout({ children }) {
+  return (
+    <html>
+      <head>
+        <title>after</title>
+      </head>
+      <body>{children}</body>
+    </html>
+  )
+}
