+++ test/e2e/app-dir/next-after-app/app/nodejs/provided-request-context/page.js
@@ -1,0 +1,9 @@
+import { unstable_after as after } from 'next/server'
+import { cliLog } from '../../../utils/log'
+
+export default function Page() {
+  after(() => {
+    cliLog({ source: '[page] /provided-request-context' })
+  })
+  return <div>provided-request-context</div>
+}
