+++ test/e2e/app-dir/next-after-app/index.test.ts
@@ -1,4 +1,5 @@
 /* eslint-env jest */
+/* eslint-disable jest/no-standalone-expect */
 import { nextTestSetup } from 'e2e-utils'
 import { retry } from 'next-test-utils'
 import { createProxyServer } from 'next/experimental/testmode/proxy'
@@ -25,36 +26,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   })
 
   if (skipped) return
-
-  {
-    const originalContents: Record<string, string> = {}
-
-    beforeAll(async () => {
-      const placeholder = `// export const runtime = 'REPLACE_ME'`
-
-      const filesToPatch = ['app/layout.js', 'app/route/route.js']
-
-      for (const file of filesToPatch) {
-        await next.patchFile(file, (contents) => {
-          if (!contents.includes(placeholder)) {
-            throw new Error(`Placeholder "${placeholder}" not found in ${file}`)
-          }
-          originalContents[file] = contents
-
-          return contents.replace(
-            placeholder,
-            `export const runtime = '${runtimeValue}'`
-          )
-        })
-      }
-    })
-
-    afterAll(async () => {
-      for (const [file, contents] of Object.entries(originalContents)) {
-        await next.patchFile(file, contents)
-      }
-    })
-  }
+  const pathPrefix = '/' + runtimeValue
 
   let currentCliOutputIndex = 0
   beforeEach(() => {
@@ -70,7 +42,8 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   }
 
   it('runs in dynamic pages', async () => {
-    await next.render('/123/dynamic')
+    const response = await next.fetch(pathPrefix + '/123/dynamic')
+    expect(response.status).toBe(200)
     await retry(() => {
       expect(getLogs()).toContainEqual({ source: '[layout] /[id]' })
       expect(getLogs()).toContainEqual({
@@ -85,7 +58,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   })
 
   it('runs in dynamic route handlers', async () => {
-    const res = await next.fetch('/route')
+    const res = await next.fetch(pathPrefix + '/route')
     expect(res.status).toBe(200)
     await retry(() => {
       expect(getLogs()).toContainEqual({ source: '[route handler] /route' })
@@ -93,7 +66,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   })
 
   it('runs in server actions', async () => {
-    const browser = await next.browser('/123/with-action')
+    const browser = await next.browser(pathPrefix + '/123/with-action')
     expect(getLogs()).toContainEqual({
       source: '[layout] /[id]',
     })
@@ -114,7 +87,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   })
 
   it('runs callbacks from nested unstable_after calls', async () => {
-    await next.browser('/nested-after')
+    await next.browser(pathPrefix + '/nested-after')
 
     await retry(() => {
       for (const id of [1, 2, 3]) {
@@ -131,7 +104,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
 
   describe('interrupted RSC renders', () => {
     it('runs callbacks if redirect() was called', async () => {
-      await next.browser('/interrupted/calls-redirect')
+      await next.browser(pathPrefix + '/interrupted/calls-redirect')
 
       await retry(() => {
         expect(getLogs()).toContainEqual({
@@ -144,14 +117,14 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
     })
 
     it('runs callbacks if notFound() was called', async () => {
-      await next.browser('/interrupted/calls-not-found')
+      await next.browser(pathPrefix + '/interrupted/calls-not-found')
       expect(getLogs()).toContainEqual({
         source: '[page] /interrupted/calls-not-found',
       })
     })
 
     it('runs callbacks if a user error was thrown in the RSC render', async () => {
-      await next.browser('/interrupted/throws-error')
+      await next.browser(pathPrefix + '/interrupted/throws-error')
       expect(getLogs()).toContainEqual({
         source: '[page] /interrupted/throws-error',
       })
@@ -161,7 +134,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   it('runs in middleware', async () => {
     const requestId = `${Date.now()}`
     const res = await next.fetch(
-      `/middleware/redirect-source?requestId=${requestId}`,
+      pathPrefix + `/middleware/redirect-source?requestId=${requestId}`,
       {
         redirect: 'follow',
         headers: {
@@ -183,10 +156,19 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   if (!isNextDeploy) {
     it('only runs callbacks after the response is fully sent', async () => {
       const pageStartedFetching = promiseWithResolvers<void>()
+      pageStartedFetching.promise.catch(() => {})
       const shouldSendResponse = promiseWithResolvers<void>()
+      shouldSendResponse.promise.catch(() => {})
+
       const abort = (error: Error) => {
-        pageStartedFetching.reject(error)
-        shouldSendResponse.reject(error)
+        pageStartedFetching.reject(
+          new Error('pageStartedFetching was aborted', { cause: error })
+        )
+        shouldSendResponse.reject(
+          new Error('shouldSendResponse was aborted', {
+            cause: error,
+          })
+        )
       }
 
       const proxyServer = await createProxyServer({
@@ -200,25 +182,30 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
       })
 
       try {
-        const pendingReq = next.fetch('/delay', {
-          headers: { 'Next-Test-Proxy-Port': String(proxyServer.port) },
-        })
-
-        pendingReq.then(
-          async (res) => {
-            if (res.status !== 200) {
-              const msg = `Got non-200 response (${res.status}), aborting`
-              console.error(msg + '\n', await res.text())
-              abort(new Error(msg))
+        const pendingReq = next
+          .fetch(pathPrefix + '/delay', {
+            headers: { 'Next-Test-Proxy-Port': String(proxyServer.port) },
+          })
+          .then(
+            async (res) => {
+              if (res.status !== 200) {
+                const err = new Error(
+                  `Got non-200 response (${res.status}) for ${res.url}, aborting`
+                )
+                abort(err)
+                throw err
+              }
+              return res
+            },
+            (err) => {
+              abort(err)
+              throw err
             }
-          },
-          (err) => {
-            abort(err)
-          }
-        )
+          )
 
         await Promise.race([
           pageStartedFetching.promise,
+          pendingReq, // if the page throws before it starts fetching, we want to catch that
           timeoutPromise(
             10_000,
             'Timeout while waiting for the page to call fetch'
@@ -253,7 +240,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
   }
 
   it('runs in generateMetadata()', async () => {
-    await next.browser('/123/with-metadata')
+    await next.browser(pathPrefix + '/123/with-metadata')
     expect(getLogs()).toContainEqual({
       source: '[metadata] /[id]/with-metadata',
       value: '123',
@@ -264,7 +251,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
     const EXPECTED_ERROR =
       /An error occurred in a function passed to `unstable_after\(\)`: .+?: Cookies can only be modified in a Server Action or Route Handler\./
 
-    const browser = await next.browser('/123/setting-cookies')
+    const browser = await next.browser(pathPrefix + '/123/setting-cookies')
     // after() from render
     expect(next.cliOutput).toMatch(EXPECTED_ERROR)
 
@@ -302,7 +289,7 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
           `,
         ],
       ]),
-      '/provided-request-context'
+      pathPrefix + '/provided-request-context'
     )
 
     try {
@@ -332,20 +319,21 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
             next,
             new Map([
               [
-                'app/static/page.js',
-                (await next.readFile('app/static/page.js')).replace(
-                  `// export const dynamic = 'REPLACE_ME'`,
-                  `export const dynamic = '${dynamicValue}'`
-                ),
+                `app/${runtimeValue}/static/page.js`,
+                (contents) =>
+                  contents.replace(
+                    `// export const dynamic = 'REPLACE_ME'`,
+                    `export const dynamic = '${dynamicValue}'`
+                  ),
               ],
             ]),
-            '/static'
+            pathPrefix + '/static'
           )
 
           try {
             await session.assertHasRedbox()
             expect(await session.getRedboxDescription()).toContain(
-              `Route /static with \`dynamic = "${dynamicValue}"\` couldn't be rendered statically because it used \`unstable_after\``
+              `Route ${pathPrefix}/static with \`dynamic = "${dynamicValue}"\` couldn't be rendered statically because it used \`unstable_after\``
             )
             expect(getLogs()).toHaveLength(0)
           } finally {
@@ -359,14 +347,11 @@ describe.each(runtimes)('unstable_after() in %s runtime', (runtimeValue) => {
           next,
           new Map([
             [
-              'app/invalid-in-client/page.js',
-              (await next.readFile('app/invalid-in-client/page.js')).replace(
-                `// 'use client'`,
-                `'use client'`
-              ),
+              `app/${runtimeValue}/invalid-in-client/page.js`,
+              (contents) => contents.replace(`// 'use client'`, `'use client'`),
             ],
           ]),
-          '/invalid-in-client'
+          pathPrefix + '/invalid-in-client'
         )
         try {
           await session.assertHasRedbox()
